% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aeronautes_saxatalis.tif}
\alias{Aeronautes_saxatalis.tif}
\title{Example SDM output for \if{html}{\out{<i>}}Aeronautes saxatalis\if{html}{\out{</i>}}.}
\format{
A single-layer GeoTIFF raster with the following properties:
\describe{
\item{CRS}{+proj=laea +lat_0=16.72 +lon_0=-98.44 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs}
\item{Resolution}{44950, 46150m (~50 km)}
\item{Extent}{-2073193, 3365757, -1199002, 4431298}
\item{Values}{Probablity of occurrence for Aeronautes saxatalis in continuous and binarized ("PA") probabilities, for breeding and nonbreeding seasons}
}
}
\source{
Derived from a model generated for Campbell & Belitz et al.; data from BirdLife International and Handbook of the Birds of the World (2020).
}
\description{
A GeoTIFF raster containing probability of occurrence model predictions for \if{html}{\out{<i>}}Aeronautes saxatalis\if{html}{\out{</i>}}.
The layers reflect continuous probabilities for breeding and nonbreeding seasons, then binarized (presence/absence, or "PA") probabilties for breeding and nonbreeding seasons.
}
