% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaOfExpectedOccurrence.R
\name{areaOfExpectedOccurrence}
\alias{areaOfExpectedOccurrence}
\title{Calculate the area of a region of expected occurrence from a presence/absence surface}
\usage{
areaOfExpectedOccurrence(surface, unit = "km")
}
\arguments{
\item{surface}{rasterLayer with values of NA, 0, and 1}
}
\value{
A units object of area of expected occurrence
}
\description{
This function calculates the geographic area of the occupied region of a
presence-absence model. The input must be a rasterLayer object with values
of NA (out of range), 0 (not expected to occur), or 1 (expected to occur).
}
