% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usingOccurrenceData.R
\name{calculateChangeAtPoints}
\alias{calculateChangeAtPoints}
\title{Calculate the change in rasterLayer values between rast1 and rast2 at the presence points used to create each of those surfaces.}
\usage{
calculateChangeAtPoints(rast1, rast2, rastnames, pts1, pts2, ...)
}
\arguments{
\item{rast1}{First input rasterLayer}

\item{rast2}{Second input rasterLayer}

\item{rastnames}{Character vector of layerNames to be compared (defaults to "Summer" and "Winter" for rast1 and rast2 respectively)}

\item{pts1}{Points used to create rast1}

\item{pts2}{Points used to create rast2}

\item{...}{Optional extra columns to include in output data.frame.}
}
\value{
A dataframe containing occurrence points categorized by their associated rasterLayer and the difference at that point between the associated and the opposing layer.
}
\description{
Calculate the change in rasterLayer values between rast1 and rast2 at the presence points used to create each of those surfaces.
}
