% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordsToConvexHull.R
\name{coordsToConvexHull}
\alias{coordsToConvexHull}
\title{Calculate the convex hull of occurrence coordinates}
\usage{
coordsToConvexHull(coords, proj4string)
}
\arguments{
\item{coords}{Two-column data.frame, matrix, or SpatialPoints object containing longitude and latitude}

\item{proj4string}{Character string containing CRS information for coords.}
}
\value{
An sf polygon object of specified proj4string
}
\description{
Convert occurrence coordinates to convex hull. Returned in same projection.
}
