% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordsToKDEPolygon.R
\name{coordsToKDEPolygon}
\alias{coordsToKDEPolygon}
\title{Determine the KDE of occurrence coordinates and convert to polygon at density threshold}
\usage{
coordsToKDEPolygon(coords, percentile = 50, proj4string)
}
\arguments{
\item{coords}{Two-column data.frame, matrix, or SpatialPoints object containing longitude and latitude}

\item{proj4string}{Character string containing CRS information for coords.}

\item{pecentile}{Numeric density threshold percentile at which to draw output polygon}
}
\value{
An sf polygon object of specified proj4string
}
\description{
Calculate the kernel density estimate (KDE) of input coordinates. Convert to
an sf polygon at a specified density threshold.
}
