% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataEllipse.R
\name{makeDataEllipse}
\alias{makeDataEllipse}
\title{Function to return data ellipse around a give proportion of points
(prop_points) sampled from a surface.}
\usage{
makeDataEllipse(coords, crsString, weights, n = 10000, prop_points)
}
\arguments{
\item{coords}{coordinates of a surface-as-dataframe}

\item{crsString}{string of crs coordinates, should match raster projections.}

\item{weights}{the value of the surface-as-dataframe, weights the subsampling.}

\item{n}{number of sampling replications}

\item{prop_points}{what proportion of the subsampled cells should be included in the ellipses}
}
\value{
Returns a list of length 2. First item is a spatialPolygon, the second an sf object.
}
\description{
Function to return data ellipse around a give proportion of points
(prop_points) sampled from a surface.
}
