% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractStatistics.R
\name{extractStatistics}
\alias{extractStatistics}
\title{Extract comparative statistics from two rasterlayer surfaces.}
\usage{
extractStatistics(
  rast1,
  rast2,
  rastnames = c("Summer", "Winter"),
  species,
  myEllipses = NULL,
  csvSavePath = FALSE,
  prop_points = 0.5,
  ...
)
}
\arguments{
\item{rast1}{First input rasterLayer}

\item{rast2}{Second input rasterLayer}

\item{rastnames}{Character vector of layerNames to be compared (defaults to "Summer" and "Winter" for rast1 and rast2 respectively)}

\item{species}{Character vector of species, name of subdirectory where results are (optionally) saved.}

\item{myEllipses}{List output of 'makeDataEllipse' function. If not provided, calculations to fit dataEllipse will be rerun. Arguments for makeDataEllipse function should then be provided.}

\item{csvSavePath}{Optional file path to save csv object to.}

\item{prop_points}{Proportion of the points sampled from the surface to be included in ellipses drawn}

\item{...}{Any additional arguments to be included in the output data.frame}
}
\value{
A data.frame containing species name, layerNames to be compared (defaults to "Summer" and "Winter" for rast1 and rast2 respectively), metrics of surface equivalency, ellipse areas, ratio of ellipse areas, percent of ellipse intersected, centroid latitude and longitude for each ellipse, and distance and bearing between each centroid.
}
\description{
Extract metrics of surface equivalency, ellipse overlap, ellipse centroid bearing and distance.
}
\details{
RasterLayers must have identical resolutions and extents. Schoener's D assumes
surfaces each sum to 1.
}
\seealso{
makeDataEllipse

surfaceEquivalency
}
