% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCompoundPlot.R
\name{makeCompoundPlot}
\alias{makeCompoundPlot}
\title{Make overall summary plot from two rasterlayer surfaces.}
\usage{
makeCompoundPlot(
  rast1,
  rast2,
  rastnames = c("Summer", "Winter"),
  myEllipses,
  species,
  summaryTable = NULL,
  pngSaveDirectory = FALSE,
  ...
)
}
\arguments{
\item{rast1}{First input rasterLayer}

\item{rast2}{Second input rasterLayer}

\item{rastnames}{Character vector of layerNames to be compared (defaults to "Summer" and "Winter" for rast1 and rast2 respectively)}

\item{myEllipses}{List output of 'makeDataEllipse' function. If not provided, calculations to fit dataEllipse will be rerun. Arguments for makeDataEllipse function should then be provided.}

\item{species}{Character vector of species, name of subdirectory where results are (optionally) saved.}

\item{summaryTable}{Optional table of values to include under the plot. See output of 'extractStatistics'}

\item{pngSaveDirectory}{Optional file path to save plot object to.}

\item{...}{DataFrame column nams to populated "geom_tile(aes())", e.g., : x=x,y=y,fill=value}
}
\value{
A list containing species name, rastname, metrics of surface equivalency, ellipse areas, ratio of ellipse areas, percent of ellipse intersected, centroid latitude and longitude for each ellipse, and distance and bearing between each centroid.
}
\description{
Make overall summary plot from two rasterlayer surfaces.
}
\seealso{
makeDataEllipse

surfaceEquivalency

extractStatistics
}
