% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataEllipse.R
\name{makeEllipses}
\alias{makeEllipses}
\title{makeEllipses}
\usage{
makeEllipses(rast1, rast2)
}
\arguments{
\item{rast1}{First input rasterLayer}

\item{rast2}{Second input rasterLayer}
}
\description{
Wrapper around makeDataEllipse to automatically call the correct columns.
Assumes columns are arranged as x,y,value.
}
