% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSeasonPlot.R
\name{makePlot}
\alias{makePlot}
\title{Make a seasonal plot, save to a specific path.}
\usage{
makePlot(
  surface,
  species,
  season,
  maxCol = NULL,
  coordArgs = NULL,
  savePath = NULL,
  ...
)
}
\arguments{
\item{surface}{array data presented in a 3-column dataframe}

\item{species}{species name (character) to include in title and file path}

\item{season}{season name (character) to specify plot color and file path. Defaults to muted red for "Summer" and muted blue for other seasons.}

\item{maxCol}{color (character) to overwrite season argument, sets maximum color.}

\item{coordArgs}{optional coordinate arguments to populate coord_sf.}

\item{savePath}{optional (character) file path to save plot.}

\item{...}{Aesthetic arguments to populate ggplot2::geom_tile(). Suggested use might include "x=x,y=y,fill=value" or similar, depending on the column names of the 'surface' parameter.}
}
\value{
a ggplot object
}
\description{
Make a seasonal plot, save to a specific path.
}
\seealso{
surface2df
}
