% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsFromPolygons.R
\name{statsFromPolygons}
\alias{statsFromPolygons}
\title{Extract statistics from two sf polygons.}
\usage{
statsFromPolygons(sf1, sf2, ...)
}
\arguments{
\item{sf1}{sfc_MULTIPOLYGON representing some surface from time 1}

\item{sf2}{sfc_MULTIPOLYGON representing some surface from time 2}

\item{...}{Additional arguments to append to the output data.frame}
}
\value{
A data.frame containing comparison metrics between time 1 and 2.
}
\description{
Extract metrics of range size change, seasonality, and centroid distance
from two sf polygons.
}
