% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usingOccurrenceData.R
\name{getOccDat}
\alias{getOccDat}
\alias{getSeasonalChangeAtSampleSites}
\title{Get occurrence data and examine change across seasons (or other time periods) at those points.}
\usage{
getOccDat(species, occDatPath, season = season)

getSeasonalChangeAtSampleSites(
  rast1,
  rast2,
  rastnames = c("Summer", "Winter"),
  species,
  speciesOccPts,
  saveDir
)
}
\arguments{
\item{species}{Character vector of species, name of subdirectory where occurrence records are stored.}

\item{occDatPath}{Directory location where occurrence data are stored (in subdirectories with species names).}

\item{season}{Character vector to name the new column indicating time period (defaults to 'season').}

\item{rast1}{First input rasterLayer}

\item{rast2}{Second input rasterLayer}

\item{rastnames}{Character vector of layerNames to be compared (defaults to "Summer" and "Winter" for rast1 and rast2 respectively)}

\item{speciesOccPts}{Occurrence points used in the model. In this example, requires column 'season'. Output of getOccDat function fits here.}

\item{saveDir}{Directory in which to save output.}
}
\value{
A dataframe containing occurence points chategorized into "Summer" and "Winter" months.

A dataframe containing occurence points chategorized into "Summer" and "Winter" months.
}
\description{
Get occurrence data and examine change across seasons (or other time periods) at those points.
}
