% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPAstatistics.R
\name{extractPAstatistics}
\alias{extractPAstatistics}
\title{Extract comparative statistics from two presence/absence RasterLayers.}
\usage{
extractPAstatistics(PA1, PA2, ..., appendPA = TRUE)
}
\arguments{
\item{PA1}{Presence/absence RasterLayer from season 1}

\item{PA2}{Presence/absence RasterLayer from season 2}

\item{...}{Additional arguments to append to the returned data.frame}

\item{appendPA}{True/false to paste character string "PA_" to the column name of the output metrics.}
}
\value{
A data.frame containing presence-absence surface based metrics
}
\description{
Extract metrics of range size change, seasonality, and centroid distance
from two presence/absence surfaces. Such surfaces are often generated by
thresholding a continuous probability-of-origin surface for conversion to a
binary surface.
}
\details{
RasterLayers must have identical resolutions and extents.
}
